Fix default settings log output
Remove the prefix from table names in the SQL commands saved to the log
file for users to apply as a patch.

Patches entered through the admin tools pass through the query factory
which prepends the prefix to table names. Pasting the output of from
the log file into that interface results in the table names having a
double prefix, and therefore not being valid table names.

Resolves zcadditions/ZCA-Responsive-Classic-for-Zen-Cart-1.5.6-v1.0.0#2

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>